+++ php/tests/Application/Command/DeduceTypesTest/StaticMethodCallFqcnLeadingSlash.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace A;
+
+class B
+{
+    /**
+     * @return self
+     */
+    public static function doStatic()
+    {
+
+    }
+}
+
+$data = \A\B::doStatic();
+// <MARKER>